Add integer point class

+++ src/2geom/coord.h
@@ -57,7 +57,7 @@ inline Coord infinity() {  return std::numeric_limits<Coord>::infinity();  }
 inline bool are_near(Coord a, Coord b, double eps=EPSILON) { return a-b <= eps && a-b >= -eps; }
 inline bool rel_error_bound(Coord a, Coord b, double eps=EPSILON) { return a <= eps*b && a >= -eps*b; }
 
-typedef long IntCoord;
+typedef int IntCoord;
 
 } // end namespace Geom
